Faster version of hash join - addressing code review comments
