+++ cpp/src/arrow/compute/exec/hash_join.h
@@ -36,74 +36,6 @@ namespace compute {
 
 using arrow::util::AccumulationQueue;
 
-class ARROW_EXPORT HashJoinSchema {
- public:
-  Status Init(JoinType join_type, const Schema& left_schema,
-              const std::vector<FieldRef>& left_keys, const Schema& right_schema,
-              const std::vector<FieldRef>& right_keys, const Expression& filter,
-              const std::string& left_field_name_prefix,
-              const std::string& right_field_name_prefix);
-
-  Status Init(JoinType join_type, const Schema& left_schema,
-              const std::vector<FieldRef>& left_keys,
-              const std::vector<FieldRef>& left_output, const Schema& right_schema,
-              const std::vector<FieldRef>& right_keys,
-              const std::vector<FieldRef>& right_output, const Expression& filter,
-              const std::string& left_field_name_prefix,
-              const std::string& right_field_name_prefix);
-
-  static Status ValidateSchemas(JoinType join_type, const Schema& left_schema,
-                                const std::vector<FieldRef>& left_keys,
-                                const std::vector<FieldRef>& left_output,
-                                const Schema& right_schema,
-                                const std::vector<FieldRef>& right_keys,
-                                const std::vector<FieldRef>& right_output,
-                                const std::string& left_field_name_prefix,
-                                const std::string& right_field_name_prefix);
-
-  bool HasDictionaries() const;
-
-  bool HasLargeBinary() const;
-
-  Result<Expression> BindFilter(Expression filter, const Schema& left_schema,
-                                const Schema& right_schema, ExecContext* exec_context);
-  std::shared_ptr<Schema> MakeOutputSchema(const std::string& left_field_name_suffix,
-                                           const std::string& right_field_name_suffix);
-
-  bool LeftPayloadIsEmpty() { return PayloadIsEmpty(0); }
-
-  bool RightPayloadIsEmpty() { return PayloadIsEmpty(1); }
-
-  static int kMissingField() {
-    return SchemaProjectionMaps<HashJoinProjection>::kMissingField;
-  }
-
-  SchemaProjectionMaps<HashJoinProjection> proj_maps[2];
-
- private:
-  static bool IsTypeSupported(const DataType& type);
-
-  Status CollectFilterColumns(std::vector<FieldRef>& left_filter,
-                              std::vector<FieldRef>& right_filter,
-                              const Expression& filter, const Schema& left_schema,
-                              const Schema& right_schema);
-
-  Expression RewriteFilterToUseFilterSchema(int right_filter_offset,
-                                            const SchemaProjectionMap& left_to_filter,
-                                            const SchemaProjectionMap& right_to_filter,
-                                            const Expression& filter);
-
-  bool PayloadIsEmpty(int side) {
-    ARROW_DCHECK(side == 0 || side == 1);
-    return proj_maps[side].num_cols(HashJoinProjection::PAYLOAD) == 0;
-  }
-
-  static Result<std::vector<FieldRef>> ComputePayload(const Schema& schema,
-                                                      const std::vector<FieldRef>& output,
-                                                      const std::vector<FieldRef>& filter,
-                                                      const std::vector<FieldRef>& key);
-};
-
 class HashJoinImpl {
  public:
   using OutputBatchCallback = std::function<void(int64_t, ExecBatch)>;
@@ -123,6 +55,7 @@ class HashJoinImpl {
   virtual Status ProbeSingleBatch(size_t thread_index, ExecBatch batch) = 0;
   virtual Status ProbingFinished(size_t thread_index) = 0;
   virtual void Abort(TaskScheduler::AbortContinuationImpl pos_abort_callback) = 0;
+  virtual std::string ToString() const = 0;
 
   static Result<std::unique_ptr<HashJoinImpl>> MakeBasic();
   static Result<std::unique_ptr<HashJoinImpl>> MakeSwiss();
